checkpoint examples
