+++ .gitignore
@@ -7,3 +7,5 @@
 /test/Manifest.toml
 
 .DS_Store
+
+docs/imgs/
